Refactor templatesToRender, add tooltips, fix wording
